Export `std::io` instead of always using thin reimplementation
This allows most implementers of `BinRead` to use std features
on the reader instead of restricting them to the subset of
features exposed by the no_std reimplementation. This also
appears to have been the original intent of the `binread::io`
module, per the module-level comment.

This is a breaking change because `iter_bytes` is called `bytes`
in std and consumes the reader, so this has been changed in the
binread implementation to conform to that API.

This commit also fixes the earlier binread implementations not
handling errors as std does; this is now fixed (by copying
from std).

(Hopefully rust-lang/rust#48331 will be addressed someday and then
this code can disappear entirely.)